requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
httpx==0.18.2

python-dotenv==0.19.2

psycopg2-binary==2.9.3
databases==0.6.0
SQLAlchemy==1.4.39
asyncpg==0.25.0
aiosqlite==0.17.0
alembic==1.7.5

construct-typing==0.5.2
pydantic==1.9.1

# auth and encryption
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.5

starlette~=0.19.1
pytest~=7.1.2
yarl~=1.7.2
